+++ android/src/main/java/com/learnium/RNDeviceInfo/RNDeviceModule.java
@@ -12,6 +12,7 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.FeatureInfo;
 import android.content.res.Configuration;
+import android.location.LocationManager;
 import android.net.wifi.WifiManager;
 import android.net.wifi.WifiInfo;
 import android.os.Build;
@@ -23,6 +24,7 @@ import android.view.WindowManager;
 import android.webkit.WebSettings;
 import android.telephony.TelephonyManager;
 import android.text.format.Formatter;
+import android.text.TextUtils;
 import android.app.ActivityManager;
 import android.util.DisplayMetrics;
 import android.hardware.Camera;
@@ -36,6 +38,7 @@ import com.facebook.react.bridge.ReactMethod;
 import com.facebook.react.bridge.Callback;
 import com.facebook.react.bridge.Promise;
 import com.facebook.react.bridge.WritableArray;
+import com.facebook.react.bridge.WritableMap;
 
 import java.util.Collections;
 import java.util.HashMap;
@@ -356,6 +359,37 @@ public class RNDeviceModule extends ReactContextBaseJavaModule {
     }
 
     p.resolve(promiseArray);
+  }
+
+  @ReactMethod
+  public void isLocationEnabled(Promise p) {
+      boolean locationEnabled = false;
+
+      if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
+        LocationManager mLocationManager = (LocationManager) reactContext.getApplicationContext().getSystemService(Context.LOCATION_SERVICE);
+        locationEnabled = mLocationManager.isLocationEnabled();
+      } else if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
+        int locationMode = Settings.Secure.getInt(reactContext.getContentResolver(), Settings.Secure.LOCATION_MODE, Settings.Secure.LOCATION_MODE_OFF);
+        locationEnabled = locationMode != Settings.Secure.LOCATION_MODE_OFF;
+      } else {
+        String locationProviders = Settings.Secure.getString(reactContext.getContentResolver(), Settings.Secure.LOCATION_PROVIDERS_ALLOWED);
+        locationEnabled = !TextUtils.isEmpty(locationProviders);
+      }
+
+      p.resolve(locationEnabled);
+  }
+
+  @ReactMethod
+  public void getAvailableLocationProviders(Promise p) {
+    LocationManager mLocationManager = (LocationManager) reactContext.getApplicationContext().getSystemService(Context.LOCATION_SERVICE);
+    final List<String> providers = mLocationManager.getProviders(false);
+
+    WritableMap providersAvailability = Arguments.createMap();
+    for (String provider : providers) {
+      providersAvailability.putBoolean(provider, mLocationManager.isProviderEnabled(provider));
+    }
+
+    p.resolve(providersAvailability);
   }
 
   public String getInstallReferrer() {