+++ default/index.js
@@ -46,5 +46,7 @@ module.exports = {
   getPowerState: () => Promise.resolve({}),
   hasSystemFeature: () => Promise.resolve(false),
   getSystemAvailableFeatures: () => Promise.resolve([]),
-  getCameraPresence: () => Promise.resolve(false)
+  getCameraPresence: () => Promise.resolve(false),
+  isLocationEnabled: () => Promise.resolve(false),
+  getAvailableLocationProviders: () => Promise.resolve({}),
 };