+++ deviceinfo.js
@@ -375,5 +375,11 @@ export default {
   },
   getSystemAvailableFeatures: function() {
     return RNDeviceInfo.getSystemAvailableFeatures();
-  }
+  },
+  isLocationEnabled: function() {
+    return RNDeviceInfo.isLocationEnabled();
+  },
+  getAvailableLocationProviders: function() {
+    return RNDeviceInfo.getAvailableLocationProviders();
+  },
 };