+++ example/App.js
@@ -81,6 +81,8 @@ export default class App extends Component<Props> {
       deviceJSON.hasSystemFeature = ios ? false : await DeviceInfo.hasSystemFeature('amazon.hardware.fire_tv');
       deviceJSON.getSystemAvailableFeatures = ios ? [] : await DeviceInfo.getSystemAvailableFeatures();
       deviceJSON.powerState = ios ? await DeviceInfo.getPowerState() : '';
+      deviceJSON.isLocationEnabled = await DeviceInfo.isLocationEnabled();
+      deviceJSON.getAvailableLocationProviders = await DeviceInfo.getAvailableLocationProviders();
     } catch (e) {
       console.log('Trouble getting device info ', e);
     }